Homework 2: Work_history Merged branches Homework completed

+++ src/Git_and_java_basics/Work_history/src/main/java/org/example/SalesManager.java
@@ -17,18 +17,18 @@ public class SalesManager {
         return max;
     }
 
-    protected int min() {
-        int min = Integer.MAX_VALUE;
-        for (int sale: sales) {
+    protected long min() {
+        long min = Integer.MAX_VALUE;
+        for (long sale: sales) {
             if (sale < min) {
                 min = sale;
             }
         }
         return min;
     }
-    public int croppedAverage() {
-        int sum = 0;
-        for (int sale: sales) {
+    public long croppedAverage() {
+        long sum = 0;
+        for (long sale: sales) {
             sum +=sale;
         }
         return (sum - min() - max()) / (sales.length - 2);
